大货下单折扣，加入购物车可小数

+++ YXDB/Procedures/ShoppingCart/P_AddShoppingCartBatchIn.sql
@@ -18,7 +18,7 @@ CREATE PROCEDURE [dbo].[P_AddShoppingCartBatchIn]
 @OrderType int,
 @ProductDetailID nvarchar(64),
 @ProductID nvarchar(64),
-@Quantity int=1,
+@Quantity decimal(18,4)=1,
 @GUID nvarchar(64)='',
 @UserID nvarchar(64),
 @Remark nvarchar(max),
